introduce first attempt of unittest
